Split Reachability config into two implementations.
